+++ rai_test_utils/rai_test_utils/datasets/vision/__init__.py
@@ -1,10 +1,11 @@
 # Copyright (c) Microsoft Corporation
 # Licensed under the MIT License.
 
-"""Namespace for tabular datasets."""
+"""Namespace for vision datasets."""
 
-from .object_detection_data_utils import load_fridge_object_detection_dataset
+from .object_detection_data_utils import get_images, load_fridge_object_detection_dataset
 
 __all__ = [
+    "get_images",
     "load_fridge_object_detection_dataset"
 ]
