+++ rai_test_utils/requirements-object-detection.txt
@@ -1,0 +1,2 @@
+torch==1.13.1
+torchvision==0.14.1
