+++ rai_test_utils/tests/test_model_utils.py
@@ -7,13 +7,12 @@ from ml_wrappers import wrap_model
 from rai_test_utils.datasets.tabular import (create_housing_data,
                                              create_iris_data,
                                              create_simple_titanic_data)
-from rai_test_utils.datasets.vision import load_fridge_object_detection_dataset
+from rai_test_utils.datasets.vision import get_images, load_fridge_object_detection_dataset
 from rai_test_utils.models import (create_models_classification,
                                    create_models_object_detection,
                                    create_models_regression)
 from rai_test_utils.models.sklearn import \
     create_complex_classification_pipeline
-from rai_test_utils.utilities import get_images
 
 
 class TestModelUtils:
